+++ NEWS.md
@@ -6,6 +6,7 @@
 #### Highlights
 * Fix ReadObs() so date/time format works when importing as a data.table object
 * Prevent scientific notation when writing files
+* Add ReadClassData() function
 
 ## HYPEtools 1.5.1 (2023-08-24)
 *Enhancements and bug fixes*